requirements.txt
astroid==1.6.3
Django==2.0.4
django-cleanup==2.1.0
django-dynamic-preferences==1.5
django-filter==1.1.0
django-jet==1.0.7
django-webpack-loader==0.5.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
persisting-theory==0.2.1
Pillow==5.0.0
pycodestyle==2.3.1
pyflakes==1.6.0
pypugjs==5.1.3
pytz==2018.4
six==1.11.0
wrapt==1.10.11
